+++ .eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    extends: ['scratch', 'scratch/node', 'plugin:import/recommended'],
-    env: {
-        es6: true
-    }
-};