+++ commitlint.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-    extends: ['@commitlint/config-conventional'],
-    ignores: [message => message.startsWith('chore(release):')]
-};
+  extends: ['@commitlint/config-conventional'],
+  ignores: [message => message.startsWith('chore(release):')],
+}